move tokenization encode / decode to separate utility file
